RedisList._slice_to_indices() always returns range
Adhere to the Robustness Principle: Be liberal in what you accept, be
conservative in what you send.  `RedisList._slice_to_indices()` always
could accept a slice or an index, but in the former case returned a
`range` and in the latter case returned a `tuple`.  Now, it always
returns a `range`.