Flesh out RedisDeque() code and tests (#40)
* Make RedisList.__eq__() more memory efficient

Before, we were converting `RedisList`s into Python `list`s (implicitly
via a slice) and comparing the Python `list`s.  Now, iterate over the
`RedisList`s element by element and compare the pairs, reducing memory
use.

* Make Redis transactions safer

Remember to issue `MULTI` before `EXEC` to execute a group of commands
transactionally.  More info:

https://redis.io/topics/transactions

* Simplify code with yield from

* Upgrade requirements

* Fix bounds checking in RedisList.pop()

* Use only context manager to watch keys

Ditch the nonsensical and brain scrambling decorator altogether.

* Define RedisDeque.append() in terms of .extend()

Don't repeat myself.

* Upgrade requirements

* Hopefully speed up Travis CI builds...

...by testing against only Python 3.4, 3.5, and 3.6.

* Dynamically enter context managers

This removes repetition when doing things like comparing a `RedisList`
on one Redis instance to a `RedisList` on a different Redis instance.
In particular, this allows us to dynamically create and enter context
managers to `WATCH` Redis keys on multiple Redis instances.

* Make transactions safer...

…by always issuing `MULTI` before getting/setting values as the last
step of a transaction.

* Upgrade requirements

* RedisList._slice_to_indices() always returns range

Adhere to the Robustness Principle: Be liberal in what you accept, be
conservative in what you send.  `RedisList._slice_to_indices()` always
could accept a slice or an index, but in the former case returned a
`range` and in the latter case returned a `tuple`.  Now, it always
returns a `range`.

* Make transactions safer...

- Require that we be in a `WATCH`ed context before calling `_populate()`
  on any Pottery container

- Use a single entry point to `WATCH` keys: the `self._watch()` context
  manager

Among other things, this commit makes it safe to call
`RedisList.extend()` with another `RedisList` (and in this case, we
`WATCH` the keys for both `RedisList`s).

* Bump version number