Design/implement/test CachedOrderedDict (#87)
* Design/implement/test CachedOrderedDict

`CachedOrderedDict` is a crazy data structure for representing things
like search result sets.  Some of the results may be hydrated, others
may not be, and order is important.

You can use `CachedOrderedDict` as a context manager.  Upon entering,
it'll populate the hits from cache.  In the body, you can hydrate the
misses.  Upon exiting, it'll cache the misses.

* Bump version number

* Don't test for sorted imports on Python 3.5

isort is broken on Python 3.5 for no good reason. ¯\_(ツ)_/¯