Gracefully ignore KeyError on isort import
This is a workaround for `pip` 9.0.2, which is the version that runs on
Heroku.  For some reason, importing `isort` import imports `requests`,
and if we import `requests` before/without importing `pip`, we get a
`KeyError`.

We only need for this import to succeed locally and on CI (basically,
anywhere where we want to run our test suite), so it's ok for this to
fail on production.

For more info:
  https://github.com/pypa/pip/issues/5079#issuecomment-373927413