Improve pipelining and watching keys over multiple Redis instances (#240)
* Improve type annotations

* Avoid unnecessary pipeline.ping() calls

* Design/implement _ContextPipeline

By using a class-based context manager (instead of a function-based
context manager), we can conditionally do stuff or not do stuff based on
whether there's an active exception.

* Upgrade requirements

* Write abstract & concrete ._bit_offsets() methods

Inspired by the Single-Responsibility Principle.  This way, the concrete
method's responsibility is to encode the value, and the abstract
method's responsibility is to compute the hashes.  Thus, I've removed
`BloomFilterABC`'s dependency on `_Encodable`.

* Bump version number