Improve Makefile (#256)
* Improve Makefile

Prior to this PR, `$ make upgrade` upgraded _only_ the top-level
requirements specified in `requirements-to-freeze.txt`, not _all_ of the
requirements specified in `requirements.txt`.  Further, there was no
Makefile target to rebuild a clean `venv` and generate a clean
`requirements.txt` (useful for example when removing a requirement that
recursively pulled in other requirements).

1. Specify a `requirements` target to rebuild the `venv`
2. Make the `upgrade` target upgrade _all_ requirements
3. Upgrade all requirements
4. Add the Distributed Computing Trove classifier

* Remove --verbose flag from coveralls command

I didn't realize that `--verbose` means I'm going to avalanche vomit on
top of your head.