When possible, bail out early when quorum achieved
The beating heart of all consensus based distributed algorithms is to
scatter a computation across multiple nodes, then to gather their
results, then to evaluate whether quorum is achieved.

In some cases, _quorum_ requires gathering all of the nodes' results
(e.g., interrogating all nodes for a maximum value for a variable).

But in other cases, _quorum_ requires gathering only `n // 2 + 1` nodes'
results (e.g., figuring out if > 50% of nodes believe that I'm the owner
of a lock).

This commit implements the logic to "bail out" early when possible, in
the latter case.