Choose a random Redis database for unit tests
In its default configuration, Redis supports 16 databases.  So for unit
tests, chose a random database in the range [1, 15] (inclusive).  This
is because on my local development environment, I use database 0 for
permanent storage for other projects.

Choosing a random database allows us to `FLUSHDB` before and after every
unit test to ensure a clean database.