Remove I/O from .__repr__()s
When the Redis connection pool is exhausted, the Redis client throws
`ConnectionError`s.  Then the Sentry SDK tries to record the tracebacks.
If you have a `Redlock` or a `NextId` anywhere in those tracebacks, then
the Sentry SDK calls `repr()` on those objects.  And if those objects'
`.__repr__()`s try to talk to Redis, then the client throws nested
`ConnectionError`s.

So don't talk to Redis from `Redlock.__repr__()` or `NextId.__repr__()`.